SWANDROID-3636: Initial implementation of GL_EXT_texture_sRGB_decode
Implementation of sRGB decode. This has been only lightly tested but
seems to be working. The behaviour differs between v3.3 and v4.0+
according to Khronos bug 14934. v3.3 will apply the sampler parameter to
texelFetch but v4.0+ will not.

Tests follow in a separate commit.

The v4.0 hardware for this was broken. See GFXH-1558.