Add BEGL memory interface "Init"
On Android it appears system applications are linked against our EGL/Nexus
integration even if they don't actually use GL. Also on Android the BEGL
registration is done by an internal library constructor rather than the
application itself, so all applications linked against the library have
the memory interface registered even if they are not GL apps.

This means we cannot do anything which consumes resources in the memory
interface register, we need to defer it to when the v3d platform
bringup calls gmem_init(). In the DRM memory interface case that means not
opening the device (and therefore allocating kernel resources and the
pagetable memory) at the point the memory interface is registered.

refs #SWANDROID-4499

Change-Id: Ie18577fed3f972e0c8a7933ba6497b8a5c73760a
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>