SWANDROID-4657: driver: create /proc/brcm/config as read-write entry allowing proxy to pass information.
when dac_override capability is not allowed (which may be the case when selinux is enabled),
the /proc/brcm/config may fail to be written to by the proxy side which prevents proper
configuration information to be passed in on server startup and causes subsequent failure to
run the intended configuration.  to address this, we ensure the entry is created as
writable.

Signed-off-by: Pierre Couillaud <pierre@broadcom.com>